+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `libcasm-xtal` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed a bug in `xtal.make_superstructure` that overwrote coordinates with properties when properties were present.
+
+
 ## [2.0.0] - 2025-05-02
 
 ### Changed